feat:批量初始化测试执行集和对应用例

+++ weqa-client-platform-service/src/main/java/com/netease/vcloud/qa/service/auto/AutoTestTestSuitService.java
@@ -2,16 +2,20 @@ package com.netease.vcloud.qa.service.auto;
 
 import com.netease.vcloud.qa.dao.ClientAutoTestSuitBaseInfoDAO;
 import com.netease.vcloud.qa.dao.ClientAutoTestSuitRelationDAO;
+import com.netease.vcloud.qa.dao.ClientScriptTcInfoDAO;
 import com.netease.vcloud.qa.model.ClientAutoTestSuitBaseInfoDO;
 import com.netease.vcloud.qa.model.ClientAutoTestSuitRelationDO;
+import com.netease.vcloud.qa.model.ClientScriptTcInfoDO;
+import com.netease.vcloud.qa.result.view.UserInfoVO;
+import com.netease.vcloud.qa.service.auto.data.AutoTCSuitInfoDTO;
+import com.netease.vcloud.qa.service.auto.view.AutoScriptInfoVO;
+import com.netease.vcloud.qa.service.auto.view.TestSuitBaseInfoVO;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 /**
  * Created by luqiuwei@corp.netease.com
@@ -25,11 +29,16 @@ public class AutoTestTestSuitService {
     @Autowired
     private ClientAutoTestSuitRelationDAO clientAutoTestSuitRelationDAO ;
 
+    @Autowired
+    private ClientScriptTcInfoDAO clientScriptTcInfoDAO ;
+    @Autowired
+    private AutoTcScriptService autoTcScriptService ;
+
     public Long addNewTestSuit(String suitName,String creator) throws  AutoTestRunException{
         if (StringUtils.isBlank(suitName)){
             throw  new AutoTestRunException(AutoTestRunException.AUTO_TEST_PARAM_EXCEPTION) ;
         }
-        List<ClientAutoTestSuitBaseInfoDO> clientAutoTestSuitBaseInfoDOS = clientAutoTestSuitBaseInfoDAO.queryAutoTestSuitByName(suitName) ;
+        List<ClientAutoTestSuitBaseInfoDO> clientAutoTestSuitBaseInfoDOS = clientAutoTestSuitBaseInfoDAO.getAutoTestSuitByName(suitName) ;
         if (!CollectionUtils.isEmpty(clientAutoTestSuitBaseInfoDOS)) {
             //已经存在
             return clientAutoTestSuitBaseInfoDOS.get(0).getId() ;
@@ -49,6 +58,10 @@ public class AutoTestTestSuitService {
         if (suitId == null){
             throw  new AutoTestRunException(AutoTestRunException.AUTO_TEST_PARAM_EXCEPTION) ;
         }
+        ClientAutoTestSuitBaseInfoDO clientAutoTestSuitBaseInfoDO = clientAutoTestSuitBaseInfoDAO.getAutoTestSuitById(suitId);
+        if (clientAutoTestSuitBaseInfoDO == null) {
+            throw new AutoTestRunException(AutoTestRunException.AUTO_TEST_SUIT_IS_NOT_EXIST);
+        }
         clientAutoTestSuitRelationDAO.deleteClientTestRelationBySuit(suitId) ;
         if (CollectionUtils.isEmpty(scriptIdList)){
             return true ;
@@ -66,6 +79,83 @@ public class AutoTestTestSuitService {
         }else {
             throw  new AutoTestRunException(AutoTestRunException.AUTO_TEST_DB_EXCEPTION) ;
         }
+    }
+
+    public boolean addTestAndSuitRelation(Long suitId ,Long scriptId) throws AutoTestRunException {
+        if (suitId == null || scriptId == null) {
+            throw new AutoTestRunException(AutoTestRunException.AUTO_TEST_PARAM_EXCEPTION);
+        }
+        ClientAutoTestSuitBaseInfoDO clientAutoTestSuitBaseInfoDO = clientAutoTestSuitBaseInfoDAO.getAutoTestSuitById(suitId);
+        if (clientAutoTestSuitBaseInfoDO == null) {
+            throw new AutoTestRunException(AutoTestRunException.AUTO_TEST_SUIT_IS_NOT_EXIST);
+        }
+        List<ClientAutoTestSuitRelationDO> clientAutoTestSuitRelationDOList = new ArrayList<ClientAutoTestSuitRelationDO>();
+        ClientAutoTestSuitRelationDO clientAutoTestSuitRelationDO = new ClientAutoTestSuitRelationDO();
+        clientAutoTestSuitRelationDO.setSuitId(suitId);
+        clientAutoTestSuitRelationDO.setScriptId(scriptId);
+        clientAutoTestSuitRelationDOList.add(clientAutoTestSuitRelationDO);
+        int count = clientAutoTestSuitRelationDAO.patchInsertClientTestRelation(clientAutoTestSuitRelationDOList);
+        if (count > 0) {
+            return true;
+        } else {
+            throw new AutoTestRunException(AutoTestRunException.AUTO_TEST_DB_EXCEPTION);
+        }
+    }
+
+
+    public List<TestSuitBaseInfoVO> getTestSuitBaseInfo(String queryKey){
+        List<ClientAutoTestSuitBaseInfoDO> clientAutoTestSuitBaseInfoDOList = clientAutoTestSuitBaseInfoDAO.queryAutoTestSuitByName(queryKey) ;
+        List<TestSuitBaseInfoVO> testSuitBaseInfoVOList = new ArrayList<TestSuitBaseInfoVO>() ;
+        if (!CollectionUtils.isEmpty(clientAutoTestSuitBaseInfoDOList)){
+            for (ClientAutoTestSuitBaseInfoDO clientAutoTestSuitBaseInfoDO : clientAutoTestSuitBaseInfoDOList){
+                if (clientAutoTestSuitBaseInfoDO == null){
+                    continue;
+                }
+                TestSuitBaseInfoVO testSuitBaseInfoVO = new TestSuitBaseInfoVO() ;
+                testSuitBaseInfoVO.setId(clientAutoTestSuitBaseInfoDO.getId());
+                testSuitBaseInfoVO.setName(clientAutoTestSuitBaseInfoDO.getSuitName());
+                UserInfoVO userInfoVO = new UserInfoVO() ;
+                userInfoVO.setEmail(clientAutoTestSuitBaseInfoDO.getSuitOwner());
+                testSuitBaseInfoVO.setOwner(userInfoVO);
+                testSuitBaseInfoVOList.add(testSuitBaseInfoVO) ;
+            }
+        }
+        return testSuitBaseInfoVOList ;
+    }
+
+    public List<AutoScriptInfoVO> getTestSuitScriptInfo(Long suitId) throws  AutoTestRunException{
+        if (suitId == null){
+            throw new AutoTestRunException(AutoTestRunException.AUTO_TEST_PARAM_EXCEPTION) ;
+        }
+        ClientAutoTestSuitBaseInfoDO clientAutoTestSuitBaseInfoDO = clientAutoTestSuitBaseInfoDAO.getAutoTestSuitById(suitId) ;
+        if (clientAutoTestSuitBaseInfoDO==null){
+            throw new AutoTestRunException(AutoTestRunException.AUTO_TEST_SUIT_IS_NOT_EXIST) ;
+        }
+        List<ClientAutoTestSuitRelationDO> clientAutoTestSuitRelationDOList = clientAutoTestSuitRelationDAO.getAutoTestSuitRelationListBySuit(suitId) ;
+        List<AutoScriptInfoVO> autoScriptInfoVOList = new ArrayList<AutoScriptInfoVO>() ;
+        if (CollectionUtils.isEmpty(clientAutoTestSuitRelationDOList)){
+            return autoScriptInfoVOList ;
+        }
+        Set<Long> clientAutoTestScriptSet = new HashSet<Long>() ;
+        for (ClientAutoTestSuitRelationDO clientAutoTestSuitRelationDO : clientAutoTestSuitRelationDOList){
+           clientAutoTestScriptSet.add(clientAutoTestSuitRelationDO.getScriptId()) ;
+        }
+        List<ClientScriptTcInfoDO> clientScriptTcInfoDOList = clientScriptTcInfoDAO.getClientScriptSet(clientAutoTestScriptSet) ;
+        autoScriptInfoVOList = autoTcScriptService.buildAutoScriptInfoVOByDOList(clientScriptTcInfoDOList) ;
+        return autoScriptInfoVOList ;
+    }
+
+    public boolean initTestSuitScriptInfo(AutoTCSuitInfoDTO autoTCSuitInfoDTO) throws AutoTestRunException{
+        if (autoTCSuitInfoDTO == null){
+            return false ;
+        }
+        boolean result = false ;
+        Long suitId = this.addNewTestSuit(autoTCSuitInfoDTO.getSuitName(),"system") ;
+        List<Long> scriptIdList = autoTcScriptService.addScriptInfo(autoTCSuitInfoDTO.getScriptList()) ;
+        if (suitId !=null && scriptIdList!=null){
+            result = this.addOrUpdateTestAndSuitRelation(suitId,scriptIdList) ;
+        }
+        return result ;
     }
 
 }
